+++ paxos/src/config.zig
@@ -1,6 +1,6 @@
 pub const Config = struct {
-    node_id: u16,
-    cluster_size: u16,
+    node_id: u8,
+    cluster_size: u8,
     base_port: u16,
     address: [*:0]const u8,
 };
